calvin_models/requirements.txt
cmake
wheel
numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib
opencv-python
omegaconf
plotly
pyhash
pytorch-lightning==1.8.6
torch==1.13.1
torchvision
MulticoreTSNE
gitpython
scipy
sentence-transformers
setuptools==57.5.0
gym
moviepy
tqdm
termcolor
wandb
